Move code to lib.rs
